Considering additionalProperties
I encountered an issue during my use of csval that if I used the additionalProperties keyword in jsonSchema it was not being respected.  I believe I've fixed it in this PR and I updated the tests to match the new expected errors.  I also added a test to prove functionality.